moved newOrderljs into forms section in VSCode and created createEditItemsForm.js
