+++ events/navigationEvents.js
@@ -1,5 +1,5 @@
 import { getOrders, showOrders } from '../api/orderData';
-import newOrderForm from '../components/newOrderForm';
+import newOrderForm from '../forms/newOrderForm';
 
 const navigationEvents = () => {
   document.querySelector('#view-orders').addEventListener('click', () => {
